requirements.txt
pip
Flask
cassandra-driver
requests_cache
matplotlib